refactor(asset): 合并 AssetServer/ShaderAsset 到 prism-asset formats
- prism-asset: 新增 formats.rs；runtime typed/runtime/mod 接入场景与资源格式；
  cooker/scene 调整；Cargo.toml 依赖更新
- prism-engine: 删除 asset_server.rs / shader_asset.rs，asset_resolver 改接
  prism-asset；新增 app/subsystem、scene/scene_system；config/launch/engine 适配
- prism-platform: 新增 config.rs 模块，context/input/lib 与 Cargo 依赖调整
- prism-app: hook.rs、render_runner、app/lib 适配新资源通道
- prism-render: buffer/capabilities/context/forward/graph/ibl 小幅调整
- docs/DESIGN.md 同步；更新 Cargo.lock（根 + game）

+++ crates/prism-render/src/ibl.rs
@@ -1118,7 +1118,7 @@ impl IblResources {
         let brdf_mem_req = unsafe { device.get_image_memory_requirements(brdf_image) };
         let brdf_memory = allocate_device_memory(&device, mem_props, brdf_mem_req)?;
         unsafe { device.bind_image_memory(brdf_image, brdf_memory, 0) }?;
-        let brdf_staging_size = (F as u64 * F as u64 * 4) as u64;
+        let brdf_staging_size = F as u64 * F as u64 * 4;
         let (brdf_staging, brdf_staging_memory) =
             create_staging_buffer(&device, mem_props, brdf_staging_size)?;
         {
